Move Home.jsx to Pages folder (#7)
* install reduxjs/toolkit, react/redux and axios

* add stateMangement of Redux

* add folder Layout in Components

* add page Home to show Header and Footer

* provide the store with React-Redux Provider

* add component Home

* add file .env to add future enviroment variables

* reestructuring folders

* move Home to Pages folder

* some fix
